chore(handlers): create post type archive handler
