+++ packages/wp-source/src/libraries/handlers/postArchive.ts
@@ -1,34 +1,6 @@
-import { Handler } from "../../../types";
+import postTypeArchiveHandler from "./postTypeArchive";
 
-const postArchiveHandler: Handler = async ({ route, state, libraries }) => {
-  const { api, populate, parse, getTotal, getTotalPages } = libraries.source;
-  const { page, query } = parse(route);
-
-  // 2. fetch the specified page
-  const response = await api.get({
-    endpoint: state.source.postEndpoint,
-    params: { search: query.s, page, _embed: true, ...state.source.params }
-  });
-
-  // 3. populate response and add page to data
-  const items = await populate({ response, state });
-  if (page > 1 && items.length === 0)
-    throw new Error(`post archive doesn't have page ${page}`);
-
-  // 4. get posts and pages count
-  const total = getTotal(response);
-  const totalPages = getTotalPages(response);
-
-  // 5. add data to source
-  Object.assign(state.source.data[route], {
-    type: "post",
-    items,
-    total,
-    totalPages,
-    isArchive: true,
-    isPostTypeArchive: true,
-    isPostArchive: true
-  });
-};
-
-export default postArchiveHandler;
+export default postTypeArchiveHandler({
+  type: "post",
+  endpoint: "posts"
+});
